tests/wpt/web-platform-tests/web-bundle/subresource-loading/subresource-loading-non-utf8-query-encoding.https.tentative.html

<!DOCTYPE html>
<meta charset="euc-jp"/>
<title>WebBundle subresource loading with non utf-8 query encoding</title>
<link
  rel="help"
  href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md"
/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <link
  rel="webbundle"
  href="../resources/wbn/non-utf8-query-encoding.wbn"
  resources="https://web-platform.test:8444/web-bundle/resources/wbn/static-element/resources/script.js?x=��"
/>

<script id="script" src="/web-bundle/resources/wbn/static-element/resources/script.js?x=��"></script>

<script>
  const onLoadPromise = new Promise((resolve) => {
    window.addEventListener('load', resolve, false);
  });

  promise_test(async () => {
    await onLoadPromise;
    assert_equals(resources_script_result, 'loaded from webbundle');
  }, "Query encoding should be correctly handled in non utf-8 encoding document.");
</script>
</body>
